requirements.txt
torch
torchvision
opencv-python
numpy
pandas
scikit-learn
albumentations
segmentation-models-pytorch
pyyaml
tqdm
